third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/object-src-no-url-allowed-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
This test passes if there isn't a console message saying the plugin was blocked. 
